fixed ambigious mapping

+++ src/main/java/com/example/authTesting/controller/ProductPublicController.java
@@ -22,7 +22,7 @@ public class ProductPublicController {
 
     // Get product by ID (public)
     @GetMapping("/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable int id) {
+    public ResponseEntity<Product> getProductById(@PathVariable Long id) {
         Optional<Product> product = productService.getProductById(id);
         return product.map(value -> new ResponseEntity<>(value, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
